pyproject.toml
[tool.poetry]
name = "goit-web-hw-09"
version = "0.1.0"
description = ""
authors = ["R.Anatolii <tv.weplay@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
colorama = "^0.4.6"
scrapy = "^2.11.2"
requests = "^2.32.3"
mongoengine = "^0.29.1"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.5.2"
redis = "^5.1.1"
redis-lru = "^0.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
